tools/utils: move the w32 header
