+++ tools/utils/ntifs_win32.h
